requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
HTMLParser==0.0.2
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
multidict==4.7.6
olefile==0.46
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytest-httpserver==0.3.4
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
texttable==1.6.2
urllib3==1.25.9
vt==3.1.3.7
vt-py==0.5.4
wcwidth==0.2.5
Werkzeug==1.0.1
yarl==1.4.2
zipp==3.1.0
